+++ app/Http/Controllers/CollectionController.php
@@ -7,10 +7,24 @@ use App\Http\Requests\StoreCollection;
 use App\Http\Requests\UpdateCollectionRequest;
 use App\SubCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Storage;
 
 class CollectionController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth')->only(['adminCollections', 'create', 'store','edit', 'update', 'destroy', 'active']);
+    }
+
+    public function adminCollections()
+    {
+        $this->authorize('adminCollections', new Collection());
+        return view('backoffice.collections.list');
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -19,10 +33,10 @@ class CollectionController extends Controller
     public function index()
     {
         // a extarnaliser vers view composer
-        $collections = Collection::with(['categories', 'categories.subCategories', 'categories.subCategories.products'])->get();
+        /* $collections = Collection::with(['categories', 'categories.subCategories', 'categories.subCategories.products'])->get();
         return view('front.shop-left-sidebar', [
             //'collections' => $collections
-        ]);
+        ]); */
     }
 
     /**
@@ -32,6 +46,7 @@ class CollectionController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', new Collection());
         return view('backoffice.collections.create');
     }
 
@@ -43,6 +58,8 @@ class CollectionController extends Controller
      */
     public function store(StoreCollection $request)
     {
+        $this->authorize('create', new Collection());
+        
         $hasFile1 = $request->hasFile('image1');
         $hasFile2 = $request->hasFile('image2');
         if ($hasFile1) {
@@ -61,6 +78,7 @@ class CollectionController extends Controller
             'image1' => $file1_name,
             'image2' => $file2_name ?? null,
             'description' => $request->description,
+            'active' => false
         ]);
 
         session()->flash('status', 'collection add successfully');
@@ -75,8 +93,8 @@ class CollectionController extends Controller
      */
     public function show($id)
     {
-        $collection = Collection::with('categories', 'categories.subCategories')->findOrFail($id);
-        return $collection;
+        /* $collection = Collection::with('categories', 'categories.subCategories')->findOrFail($id);
+        return $collection; */
     }
 
     /**
@@ -87,6 +105,7 @@ class CollectionController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', new Collection());
         $collection = Collection::findOrFail($id);
         return view('backoffice.collections.edit', [
             'collection' => $collection
@@ -102,6 +121,7 @@ class CollectionController extends Controller
      */
     public function update(UpdateCollectionRequest $request, $id)
     {
+        $this->authorize('update', new Collection());
         $collection = Collection::findOrFail($id);
         $collection->name = $request->name;
 
@@ -139,7 +159,7 @@ class CollectionController extends Controller
      */
     public function destroy($id)
     {
-
+        $this->authorize('delete', new Collection());
         $collection = Collection::findOrFail($id);
 
         $collection->delete();
@@ -150,11 +170,25 @@ class CollectionController extends Controller
 
     public function productsByCollectionId($id)
     {
-        $collection = Collection::find($id);
+        $collection = Collection::with(['categories' => function($q){
+            $q->where('active', 1);
+        }
+        , 'categories.subCategories' => function($q){
+            $q->where('active', 1);
+        } 
+        , 'categories.subCategories.products' => function($q){
+            $q->where('active', 1);
+        }])->where('active',1)->findOrFail($id);
+        $categories = $collection->categories->filter(function($category){
+            return $category->active == 1;
+        });
+        $products = $collection->categories[0]->subCategories[0]->products()->where('active',1);
+       //dd($collection->categories[0]->subCategories);
         return view(
             'front.shop-left-sidebar',
             [
-                'products' => $collection->categories[0]->subCategories[0]->products,
+                //'products' => $collection->categories[0]->subCategories[0]->products()->paginate(1),
+                'products' => $products->paginate(Config::get('constants.options.option_product_pagination', 10)),
                 // 'collections' => Collection::all()
             ]
         );
@@ -162,6 +196,7 @@ class CollectionController extends Controller
 
     public function active(Request $request, $id)
     {
+        $this->authorize('active', new Collection());
         $collection = Collection::findOrFail($id);
         $collection->active = $request->active == 'true' ? 1 : 0;
         $collection->save();
