+++ app/Http/Controllers/OrderController.php
@@ -8,39 +8,16 @@ use App\Product;
 use App\Size;
 use App\Status;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Cache;
 
 class OrderController extends Controller
 {
 
-    public $productList;
-
     public function __construct()
     {
-        /* $this->productList =  Cache::remember('productsCardCache', now()->addSeconds(5), function () {
-            return [];
-        }); */
+       $this->middleware('auth')->only(['listNotConfirmed', 'listConfirmed', 'listInprogress', 'listCanceled', 'show', 'edit', 'update', 'destroy', 'updateStatut']);
     }
-
-
-
-
-
-  /*   public function cache(Request $request)
-    {
-        if (!$request->session()->has('productsCardSession')) {
-            $request->session()->put('productsCardSession', []);
-        }
-        $product = Product::with('sizes')->findOrFail($request->product_id);
-        if (isset($request->size)) {
-            $size = Size::findOrFail($request->size);
-            $product->setRelation('sizes', $size);
-        }
-
-        $product->quantity = $request->quantity;
-
-        $request->session()->push('productsCardSession', $product);
-    } */
 
 
 
@@ -66,6 +43,7 @@ class OrderController extends Controller
 
     public function listNotConfirmed()
     {
+        $this->authorize('listNotConfirmed', new Order());
         $orders = Order::whereHas('status', function($q){
             $q->where('label','not confirmed');
         })->get();
@@ -75,6 +53,7 @@ class OrderController extends Controller
     }
     public function listConfirmed()
     {
+        $this->authorize('listConfirmed', new Order());
         $orders = Order::whereHas('status', function($q){
             $q->where('label','confirmed');
         })->get();
@@ -84,6 +63,7 @@ class OrderController extends Controller
     }
     public function listInprogress()
     {
+        $this->authorize('listInprogress', new Order());
         $orders = Order::whereHas('status', function($q){
             $q->where('label','in progress');
         })->get();
@@ -93,6 +73,7 @@ class OrderController extends Controller
     }
     public function listCanceled()
     {
+        $this->authorize('listCanceled', new Order());
         $orders = Order::whereHas('status', function($q){
             $q->where('label','canceled');
         })->get();
@@ -109,6 +90,9 @@ class OrderController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Auth::check()) {
+            return false;
+        }
         $statut = Status::where('label', 'not confirmed')->first();
         $products = session()->get('productsCardSession');
         $order = new Order();
@@ -117,8 +101,8 @@ class OrderController extends Controller
             $total_price += (($value->new_price ?? $value->unit_price) * $value->quantity);
         }
         $order->total_price = $total_price;
-        $order->user_id = 1;
-        $order->shipping_address = "Meknes";
+        $order->user_id = Auth::user()->id;
+        $order->shipping_address = Auth::user()->address;
         $order->status_id = $statut->id;
         $order->save();
         //$productsIds_array  = [];
@@ -142,6 +126,7 @@ class OrderController extends Controller
      */
     public function show($id)
     {
+        $this->authorize('create', new Order());
         $order = Order::with('products')->findOrFail($id);
         //dd($order->products);
         return view('backoffice.orders.partials.show',[
@@ -157,6 +142,7 @@ class OrderController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', new Order());
         $order = Order::findOrFail($id);
         return view('backoffice.orders.partials.edit', [
             'order' => $order,
@@ -172,6 +158,7 @@ class OrderController extends Controller
      */
     public function update(StoreOrderRequest $request, $id)
     {
+        $this->authorize('update', new Order());
         $order = Order::findOrFail($id);
         $order->shipping_address = $request->shipping_address;
         $order->total_price = $request->total_price;
@@ -190,6 +177,7 @@ class OrderController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', new Order());
         $order = Order::findOrFail($id);
        
         $order->delete();
@@ -199,6 +187,7 @@ class OrderController extends Controller
 
     public function updateStatut(Request $request, $id)
     {
+        $this->authorize('updateStatut', new Order());
         $order = Order::findOrFail($id);
         $status = Status::where('label',$request->status)->firstOrFail();
