+++ app/Http/Controllers/ProductController.php
@@ -4,16 +4,25 @@ namespace App\Http\Controllers;
 
 use App\Category;
 use App\Collection;
+use App\Http\Requests\SearchRequest;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Product;
 use App\SubCategory;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
+use Symfony\Component\HttpFoundation\ServerBag;
 
 class ProductController extends Controller
 {
+    
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['show', 'productsBySubCategoryId', 'search']);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -21,6 +30,7 @@ class ProductController extends Controller
      */
     public function index()
     {
+        $this->authorize('index', new Product());
         $products = Product::with('subCategory', 'subCategory.category', 'subCategory.category.collection')->get();
         /*  foreach ($products as $product) {
             if($product->subCategory->category ==null)
@@ -40,6 +50,7 @@ class ProductController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', new Product());
         $subcategories = SubCategory::all();
         return view('backoffice.products.create',[
             'subcategories' => $subcategories
@@ -54,6 +65,7 @@ class ProductController extends Controller
      */
     public function store(StoreProductRequest $request)
     {
+        $this->authorize('create', new Product());
         $hasFile1 = $request->hasFile('image1');
         $hasFile2 = $request->hasFile('image2');
         if ($hasFile1) {
@@ -89,7 +101,10 @@ class ProductController extends Controller
      */
     public function show($id)
     {
-        $product = Product::findOrFail($id);
+        //$product = Product::findOrFail($id);
+        $product = Product::where('active',1)->with(['images'=> function($q){
+            $q->where('active', '=', 1);
+        }])->findOrFail($id);
         return view('front.single-product', [
             'product' => $product
         ]);
@@ -103,6 +118,7 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('update', new Product());
         $product = Product::with('subCategory', 'subCategory.category', 'subCategory.category.collection')->findOrFail($id);
         
         $subcategories = SubCategory::all();
@@ -121,6 +137,7 @@ class ProductController extends Controller
      */
     public function update(UpdateProductRequest $request, $id)
     {
+        $this->authorize('update', new Product());
         $product = Product::findOrFail($id);
         $product->name = $request->name;
         $product->unit_price = $request->unit_price;
@@ -163,6 +180,7 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', new Product());
         $product = Product::findOrFail($id);
        
         $product->delete();
@@ -172,40 +190,28 @@ class ProductController extends Controller
 
     public function productsBySubCategoryId($id)
     {
-
-
-        $subCategory = SubCategory::findOrFail($id);
-        // dd($subCategory);
-        // dd($subCategory->products()->paginate(2));
-        // a extarnaliser vers view composer
-        $collections = Collection::with(['categories', 'categories.subCategories', 'categories.subCategories.products'])->get();
-
-        return [
-            'products' => $subCategory->products,
+        $subCategory = SubCategory::with('products')->find($id);
+        $products = $subCategory->products()->where('active', 1);
+        return view('front.shop-left-sidebar',[
+            'products' => $products->paginate(Config::get('constants.options.option_product_pagination', 10)),
             'sessionProducts' => session()->get('productsCardSession')
-            //'collections'=>$collections,
-        ];
-
-        /* return view('front.shop-left-sidebar',[
-            'products' => $subCategory->products()->paginate(4),
-            'collections'=> $collections
-        ]); */
+        ]);
     }
 
-    public function search(Request $request)
+    public function search(SearchRequest $request)
     {
-        $products = DB::table('products')
-            ->where('name', 'like', '%' . $request->name . '%')
+        $products = Product::where('name', 'like', '%' . $request->product_name . '%')
             ->where('active',1)
             ->get();
-        return [
-            'products' => $products,
-            'sessionProducts' => session()->get('productsCardSession')
-        ];
+            return view('front.shop-left-sidebar',[
+                'products' => $products,
+                'sessionProducts' => session()->get('productsCardSession')
+            ]);
     }
 
     public function active(Request $request, $id)
     { 
+        $this->authorize('active', new Product());
         $product = Product::findOrFail($id);
         
         $product->active = $request->active== 'true' ? 1 : 0;
@@ -214,6 +220,7 @@ class ProductController extends Controller
 
     public function imagesByProductId($id)
     {
+        $this->authorize('imagesByProductId', new Product());
         $product = Product::with('images')->findOrFail($id);
         return view('backoffice.products.images.list',[
             'product' => $product
